requirements.txt
#conda list -e > requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
agate=1.7.1=pypi_0
asttokens=2.4.0=pyhd8ed1ab_0
attrs=23.1.0=pypi_0
babel=2.13.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
dbt-core=1.6.6=pypi_0
dbt-extractor=0.4.1=pypi_0
dbt-postgres=1.6.6=pypi_0
dbt-semantic-interfaces=0.2.2=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
greenlet=3.0.0=pypi_0
hologram=0.0.16=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
ipython=8.16.1=pyh0d859eb_0
isodate=0.6.1=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
ld_impl_linux-64=2.38=h1181459_1
leather=0.3.4=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
logbook=1.5.3=pypi_0
markupsafe=2.1.3=pypi_0
mashumaro=3.8.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
minimal-snowplow-tracker=0.0.2=pypi_0
more-itertools=8.14.0=pypi_0
msgpack=1.0.7=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2=pypi_0
numpy=1.26.1=pypi_0
openssl=3.0.11=h7f8727e_2
packaging=23.2=pypi_0
pandas=2.1.1=pypi_0
parsedatetime=2.6=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.11.2=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pip=23.3=py312h06a4308_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.24.4=pypi_0
psycopg2-binary=2.9.9=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pypi_0
pydantic=1.10.13=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
python=3.12.0=h996f2a0_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pyhd8ed1ab_1
python-slugify=8.0.1=pypi_0
pytimeparse=1.1.8=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
requests=2.31.0=pypi_0
rpds-py=0.10.6=pypi_0
setuptools=68.0.0=py312h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=2.0.22=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
text-unidecode=1.3=pypi_0
tk=8.6.12=h1ccaba5_0
traitlets=5.11.2=pyhd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=1.26.18=pypi_0
wcwidth=0.2.8=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.4.2=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
